feat: add styles for settings header
