+++ src/assets/icons/icons.tsx
@@ -139,10 +139,26 @@ export function CreateNewChatIcon() {
     </svg>
   );
 }
+export function EditIcon() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 24 24"
+      width="24px"
+      height="24px"
+      fill="grey"
+      className="icon"
+    >
+      <path
+        d="M 18 2 L 15.585938 4.4140625 L 19.585938 8.4140625 L 22 6 L 18 2 z M 14.076172 5.9238281 L 3 17 L 3 21 L 7 21 L 18.076172 9.9238281 L 14.076172 5.9238281 z"
+      />
+    </svg>
+  );
+}
 
 export function ArrowLeftIcon() {
   return (
-    <span>&#129120;</span>
+    <span className="icon">&#129120;</span>
   );
 }
