+++ src/components/ForSettingsSidebar/SettingsHeader/SettingsHeader.scss
@@ -1,0 +1,51 @@
+@import "../../../styles/variables";
+
+.settings-header {
+    height: 56px;
+    width: 100%;
+    max-width: 420px;
+
+    display: flex;
+    align-items: center;
+    justify-content: space-around;
+
+    &__title {
+        width: 50%;
+        font-size: 22px;
+        text-align: left;
+    }
+
+    &__close-button,
+    &__edit-button,
+    &__logout-button {
+        height: fit-content;
+        min-height: 25px;
+        position: relative;
+        border: none;
+        background: none;
+    
+        cursor: pointer;
+    
+        &::before {
+            z-index: 1;
+            content: "";
+            position: absolute;
+            top: -6px;
+            left: -2px;
+    
+            width: 40px;
+            height: 40px;
+            border-radius: 50%;
+            background: none;
+            transition: linear .15s;
+        }
+    
+        &:hover {
+            &::before {
+                z-index: 1;
+                transition: linear .15s;
+                background: $grey-hover;
+            }
+        }
+    }
+}
