+++ src/styles/_global.scss
@@ -1,3 +1,5 @@
+@import "./variables";
+
 .container {
   box-sizing: border-box;
 
@@ -11,8 +13,12 @@
 }
 
 .icon {
+  position: relative;
+  z-index: 3;
   width: 100%;
   max-width: 25px;
   height: fit-content;
   margin-right: 10px;
+
+  color: $grey-lite;
 }
